Refactor hotel and room group models to update category type and handle attributes as dynamic lists

+++ lib/features/hotels/data/models/hotel_model.dart
@@ -12,7 +12,7 @@ class HotelModel {
   final String hotelId;
   final String name;
   final String destination;
-  final double category;
+  final int category;
   @JsonKey(name: 'category-type')
   final String categoryType;
   final double latitude;
